Add support for `datasets`
Not quite happy with the layout of these dataset items as they become to
wide with the hash in the description area.

Also the README.md seems to have changed order but changes were never
commited.

Also includes a fix to be able to run dev server on any port with the
`PORT` environment variable.

License: MIT
Signed-off-by: Victor Bjelkholm <git@victor.earth>