Fix mid-word line breaks & right-side whitespace
Line breaking in paragraphs in cards was set to "break-all" and would
break lines in the middle of words, making reading difficult. Changing
to "break-word" tries to break between words or at hyphens, which makes
reading cards much easier.

In addition, the headers and paragraphs in cards had a maximum width set
using the mw5 class tag, which is set as 16rem in ./src/static/app.css.
However, this maximum width only spans around 3/4 of width of the
card, so there was a column of whitespace on the right hand side of the
card. Removing the "class=mw5" constraint from the headers, paragraphs,
images, and hashes in cards allows the bounding boxes for these items to
expand to the right-side margin of the card. The title, text, and image
now fill the card and dynamically adjust width if the overall window
changes width.